components/profiler/requirements.txt
flask
PyYAML
requests
matplotlib
numpy
pillow